notify: rename JsonUpdater to JsonNotifier
