+++ bots/notify/build.gradle
@@ -28,6 +28,7 @@ module {
         requires 'org.openjdk.skara.test'
         opens 'org.openjdk.skara.bots.notify' to 'org.junit.platform.commons'
         opens 'org.openjdk.skara.bots.notify.mailinglist' to 'org.junit.platform.commons'
+        opens 'org.openjdk.skara.bots.notify.json' to 'org.junit.platform.commons'
     }
 }
