+++ bots/notify/src/main/java/module-info.java
@@ -37,7 +37,7 @@ module org.openjdk.skara.bots.notify {
     uses org.openjdk.skara.bots.notify.NotifierFactory;
     provides org.openjdk.skara.bots.notify.NotifierFactory with
             org.openjdk.skara.bots.notify.issue.IssueUpdaterFactory,
-            org.openjdk.skara.bots.notify.json.JsonUpdaterFactory,
+            org.openjdk.skara.bots.notify.json.JsonNotifierFactory,
             org.openjdk.skara.bots.notify.mailinglist.MailingListNotifierFactory,
             org.openjdk.skara.bots.notify.slack.SlackUpdaterFactory;
 }
